Mask out ContextPath from URL string.
Add new utility method to isolate changes to one place.  This removes
the ContextPath from the beginning of a URL string.  Resolves several
problems with the original code:

1.  As noted in https://github.com/ocpsoft/rewrite/issues/180 , the root
context path ("/") should not be removed.
2.  All URLs should start with a /.
3.  If the context path partially matches, then it shouldn't be removed.
E.g. http://www.example.com/example should not be removed from
http://www.example.com/examples because s is not a valid start.

This method requires that the remainder after the removal start with a
/.  If it doesn't, the URL string is passed through unchanged.