Replace collection cast with unbounded wildcard cast
This makes my IDE happy, even though it basically just moves the
location of the cast.  In this location, the cast is made explicitly, as
necessary.  It is clearly intentional, whereas the original is sometimes
done unintentionally.  In the end, the real gain is a reduction in the
noise produced by the compiler.