Restore Java EE 6 specifications
This restores the Java EE 6 specification and makes a few other changes
to simplify handling.  The goal here is for versions and specifications
to be chosen at the top of rewrite-parent/pom.xml and managed from
there.  That way they can be updated quickly and easily while being sure
that versions will match.